Switch to public key provided in DER/RFC 5480 form
The DER representation according to RFC 5480 has been introduced with
OCMF and will be used on the front plate. So this representation should
be readable from the Modbus interface as well.

This updates the BSM model and public key handling accordingly. Snapshot
signatures are about to change to DER as well soon.